Refactored the queue implementation
Queue is now a bit more logical (only requires a track id to queue)
Makes the next change easier

+++ spotify/trackinfo.go
@@ -41,7 +41,7 @@ type TrackInfo struct {
 	} `json:"artists"`
 	AvailableMarkets []interface{} `json:"available_markets"`
 	DiscNumber int `json:"disc_number"`
-	DurationMs int `json:"duration_ms"`
+	DurationMs float64 `json:"duration_ms"`
 	Explicit bool `json:"explicit"`
 	ExternalIds struct {
 		Isrc string `json:"isrc"`
